backend/requirements.txt
# Web framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Cache & Message Queue
redis==5.0.1

# AI services
openai>=1.6.1
assemblyai==0.17.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.0

# WebSocket
websockets==12.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
aiofiles==23.2.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.0